+++ main.go
@@ -3,6 +3,8 @@ package main
 import (
 	"context"
 	"fmt"
+	"github.com/cloudogu/k8s-ces-control/packages/doguinteraction"
+	"github.com/cloudogu/k8s-registry-lib/dogu/local"
 	"net"
 	"os"
 
@@ -110,8 +112,18 @@ func registerServices(client clusterClient, grpcServer grpc.ServiceRegistrar) er
 		config.CurrentLokiGatewayConfig.Password,
 	)
 
-	pbLogging.RegisterDoguLogMessagesServer(grpcServer, logging.NewLoggingService(lokiLogProvider))
-	pbDoguAdministration.RegisterDoguAdministrationServer(grpcServer, doguAdministration.NewDoguAdministrationServer(client, cesReg, config.CurrentNamespace))
+	loggingService := logging.NewLoggingService(
+		lokiLogProvider,
+		cesReg,
+		doguinteraction.NewDefaultDoguInterActor(client, config.CurrentNamespace, cesReg),
+		local.NewCombinedLocalDoguRegistry(
+			client.CoreV1().ConfigMaps(config.CurrentNamespace),
+			cesReg,
+		),
+	)
+
+	pbLogging.RegisterDoguLogMessagesServer(grpcServer, loggingService)
+	pbDoguAdministration.RegisterDoguAdministrationServer(grpcServer, doguAdministration.NewDoguAdministrationServer(client, cesReg, config.CurrentNamespace, loggingService))
 	pgHealth.RegisterDoguHealthServer(grpcServer, doguHealth.NewDoguHealthService(client))
 	debugModeService := debug.NewDebugModeService(cesReg, client, config.CurrentNamespace)
 	pbMaintenance.RegisterDebugModeServer(grpcServer, debugModeService)