+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.8.0] - 2024-05-23
+### Changed
+- Update Makefiles to 9.0.5
+
+### Added
+- [#27] API GetDoguList reports current LogLevel
+- [#25] API to set log level for a dogu
+  - when the provided log level differs from the log level set in the config, a dogu restart is triggered to apply the change
+
+### Fixed
+- A bug where the healthcheck never returned the correct container state (#29) 
+
 ## [v0.7.0] - 2024-05-02
 ### Added
 - [#23] Add new query API for Dogu-Logs